Restructure SvgrComponent typing and move the custom type to typings folder
