--- custom.d.ts
@@ -1,6 +1,0 @@
-declare module '*.svg' {
-  import SvgrComponent from './src/index';
-
-  const value: SvgrComponent;
-  export default value;
-}
